src/requirements.txt
Django>=3.1.3,<4.0
psycopg2-binary>=2.8
django-jet-reboot==1.1.1
django-import-export==2.4.0
djangorestframework==3.12.1
drf-yasg==1.20.0
django-rosetta==0.9.5
flake8==3.8.4
django-debug-toolbar==3.1.1
pillow==8.0.1
celery==5.0.5
watchdog==1.0.2
redis==3.5.3
PyYAML==5.3.1
argh==0.26.2
django-cleanup==5.1.0
django-crum==0.7.9
django-object-actions==3.0.1
